Clean code is happy code
