+++ models/user.js
@@ -10,6 +10,10 @@ const userSchema = new Schema({
         type: String,
         required: true
     },
+    fullName: {
+        type: String,
+        required: false
+    },
     email: {
         type: String,
         required: true
